mqtt_to_sql/requirements.txt
absl-py==0.12.0
aio-pika==6.8.0
aiofiles==0.6.0
aiohttp==3.6.0
aiormq==3.3.1
alembic==1.4.3
APScheduler==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.7.0
beautifulsoup4==4.8.2
bidict==0.21.2
boto3==1.17.51
botocore==1.20.51
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.6.0
colorclass==2.2.0
coloredlogs==14.3
colorhash==1.0.3
configparser==5.3.0
croniter==1.0.12
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.7
dm-tree==0.1.5
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
fbmessenger==6.0.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.14
google-api-core==1.25.0
google-auth==1.24.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
greenlet==1.0.0
grpcio==1.35.0
h11==0.9.0
h5py==2.10.0
html5lib==1.1
httpcore==0.11.1
httplib2==0.19.1
httptools==0.1.1
httpx==0.15.4
humanfriendly==9.1
ibm-cloud-sdk-core==1.5.1
ibm-cos-sdk-core==2.6.0
ibm-cos-sdk-s3transfer==2.6.0
idna==2.8
importlib-metadata==3.10.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.15.1
Js2Py==0.67
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==1.4.7
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libcst==0.3.16
load-dotenv==0.1.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mattermostwrapper==2.2
multidict==4.6.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
networkx==2.5.1
num2words==0.5.10
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.1
paho-mqtt==1.6.1
pamqp==2.3.0
pika==1.2.0
Pillow==8.1.0
prompt-toolkit==2.0.10
proto-plus==1.13.0
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.19
pydot==1.4.2
pyjsparser==2.7.1
PyJWT==1.7.1
pykwalify==1.8.0
pymongo==3.10.1
PyMuPDF==1.18.14
pyparsing==2.4.6
PyPDF2==1.26.0
PyPrind==2.11.2
pyrsistent==0.17.3
PySocks==1.7.1
pyTelegramBotAPI==3.7.7
pytesseract==0.3.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-dotenv==0.21.0
python-editor==1.0.4
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2019.3
PyYAML==5.3
questionary==1.5.2
rasa==2.5.0
rasa-sdk==2.5.0
rasa-x==0.38.1
redis==3.5.3
regex==2020.2.20
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
robobrowser==0.5.3
rocketchat-API==1.15.0
rsa==4.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.3.7
sanic==20.12.3
Sanic-Cors==0.10.0.post3
sanic-jwt==1.6.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.1
scipy==1.6.2
sentry-sdk==0.14.1
six==1.14.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smmap==4.0.0
sniffio==1.2.0
soupsieve==1.9.5
spa2num==0.0.5
SpeechRecognition==3.8.1
SQLAlchemy==1.3.22
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.2
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-probability==0.11.1
tensorflow-text==2.3.0
termcolor==1.1.0
terminaltables==3.1.0
text2num==2.2.1
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.59.0
twilio==6.50.1
typeguard==2.12.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.0.0
ujson==1.35
urllib3==1.25.8
uvloop==0.14.0
waitress==1.4.4
watson-developer-cloud==2.10.1
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6
websocket-client==0.48.0
websockets==8.1
Werkzeug==1.0.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
